Version 1.2.0 adding bmp280

+++ examples/bmp280/SPI_Normal/main.cpp
@@ -32,10 +32,10 @@ int main() {
 		printf("Test Count: %u\n", counter);
 
 		// Test 1 Temperature 
-		printf("Temperature oversampling: %i\n", bmp280.readOversampling(BMP280_Sensor::DataType_e::Temperature));
+		printf("Temperature oversampling: %u\n",static_cast<uint8_t>(bmp280.readOversampling(BMP280_Sensor::DataType_e::Temperature)));
 		printf("Temperature: %f[C]\n", bmp280.readTemperature());
 		// Test 2 Pressure
-		printf("Pressure oversampling: %i\n", bmp280.readOversampling(BMP280_Sensor::DataType_e::Pressure));
+		printf("Pressure oversampling: %u\n", static_cast<uint8_t>(bmp280.readOversampling(BMP280_Sensor::DataType_e::Pressure)));
 		printf("Pressure: %f[hPa]\n", bmp280.readPressure(BMP280_Sensor::PressureUnit_e::hPa));
 
 		printf("\n\n");
